pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "financial-analyst-agent"
version = "0.1.0"
description = "Agentic AI MVP: Multi-step financial research assistant using LangGraph + Azure AI Search"
requires-python = ">=3.11"

dependencies = [
    # ── Azure ──────────────────────────────────────────────────────────────────
    # IMPORTANT: must use beta build for vector search support
    "azure-search-documents>=11.6.0b4",
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.16.0",

    # ── LangGraph / LangChain ──────────────────────────────────────────────────
    "langgraph>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.2.0",

    # ── OpenAI (Azure backend) ─────────────────────────────────────────────────
    "openai>=1.40.0",

    # ── PDF Parsing ────────────────────────────────────────────────────────────
    # PyMuPDF 1.24+ required for find_tables() API
    "pymupdf>=1.24.0",
    # Table -> Markdown conversion
    "tabulate>=0.9.0",

    # ── Text Processing ────────────────────────────────────────────────────────
    "tiktoken>=0.7.0",
    "spacy>=3.7.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",

    # ── Config / Utilities ─────────────────────────────────────────────────────
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "tenacity>=8.3.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "ipywidgets>=8.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
